+++ SourceryTests/Generating/StencilTemplateSpec.swift
@@ -44,6 +44,18 @@ class StencilTemplateSpec: QuickSpec {
                 expect(generate("{{\"helloWorld\" | replace:\"foo\",\"bar\" }}")).to(equal("helloWorld"))
             }
 
+            it("generates camel case from snake case") {
+                expect(generate("{{\"hello_world\" | snakeCaseToCamelCase}}")).to(equal("helloWorld"))
+                expect(generate("{{\"\" | snakeCaseToCamelCase}}")).to(equal(""))
+                expect(generate("{{\"hello\" | snakeCaseToCamelCase}}")).to(equal("hello"))
+            }
+
+            it("generates snake case from camel case") {
+                expect(generate("{{\"helloWorld\" | camelCaseToSnakeCase}}")).to(equal("hello_world"))
+                expect(generate("{{\"VeryLongNameHere\" | camelCaseToSnakeCase}}")).to(equal("very_long_name_here"))
+                expect(generate("{{\"\" | camelCaseToSnakeCase}}")).to(equal(""))
+                expect(generate("{{\"hello_world\" | camelCaseToSnakeCase}}")).to(equal("hello_world"))
+            }
         }
     }
 }