+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Added inline code generation
 - Added `isClosure` property to `TypeName` to detect closure types
+- Added `snakeCaseToCamelCase` filter
+- Added `camelCaseToSnakeCase` filter
 
 ### Bug Fixes